+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   sprinting highlights
 - saving actor background information
 - show the associated SPECIAL for each skill in the character sheet
+- ability to select the Thaumaturgy associated SPECIAL
 
 ### Changed